Modified the DriveRobot program to reflect the change in drive motors which now spin the opposite direction because we don't use chain drive anymore.
Fixed a problem with the EasyOpenCV programs that were programed in a way that used a phone camera instead of a webcam. For some reason it worked the first time it was programmed, but never again after that and that's been fixed.

Practice ended before I could test this, but I changed the flow of the sample autonomous program to initialize the OpenCV cam, get a reading of the rings, and close the camera before doing anything related to the vuforia camera. Hopefully this allows both functions to work without conflicting with each other.