Fixed the CameraTesting and DriveRobot programs to correctly read the vuforia targets relative to the robot as opposed to being relative to the field. I also finished the code to determine more accurately where the correct target is when the wrong one is seen. I still need to test how well this works, because half the time the value it sends back to the phone is NaN which I assume means imaginary.
I'd also like to factor in the gyroscope in the calculations, because when the robot is zipping around the field as fast as it is, the camera will likely not see the target, but when it sees it once it can remember the orientation it was in, and then calculate the approximate angle from that. I'm not sure if I'll be able to combine these two calculations in some way, but if I can I think that it would provide some of the most accurate automation I can eke out with the limited hardware I have.

Anyway, all that is to say that I added hardware configuration to test out that functionality. I can't test it at the moment on account of the battery having a whole 5 volts so the robot can't even turn on. I'll test this out tomorrow at practice, and I'll definitely be needing the field set up for it.

To do: Finalize automated target scanning,
       Factor the target scanning into automated launching through the use of one button,
       Refactor the autonomous programs so the targets are relative to the robot and not the field.
       Begin working on the autonomous program, and applying the automated launch functions in those programs.